to long

+++ src/main/java/com/edu/roy/wx/web/WxController.java
@@ -151,7 +151,7 @@ public class WxController extends BaseController {
 					WxAccessTokenBO watbo = new WxAccessTokenBO();
 					watbo.setAccessToken(accessToken);
 					watbo.setCreated(System.currentTimeMillis());
-					watbo.setExpires((int) trMap.get("expires_in"));
+					watbo.setExpires((long) trMap.get("expires_in"));
 					cache.put(Cons.WX_ACCESS_TOKEN_KEY, new CacheValue<WxAccessTokenBO>(watbo));
 				}
 			}
